Several cache enhancements
Added .backfs_version magic file, reading it spits out the date and time of BackFS's compilation.

.backfs_control magic file changes:
    - Fixed the invalidate command to return errors e.g. if the file doesn't exist in the cache.
    - Added a free_orphans command to free buckets whose parents in the map are gone.
        (a quick and dirty way to invalidate a whole directory tree is now possible (needs root tho):
            - rm -r <cache_dir>/map/some/dir
            - echo -n free_orphans >> .backfs_control
        )

Fixed an fd leak where the opendir() in backfs_opendir() was never closed.

fscache internals: added a free_bucket_mid_queue() that is the same as free_bucket() but doesn't complain if the bucket isn't the tail of the used queue.

fscache internals: added retvals for some functions