Support FUSE2 using the FUSE3 code and some wrappers.
FUSE3 changes the signature of a couple functions to take a
fuse_file_info pointer as an additional last parameter, so write
wrappers for these that just pass NULL to the real implementation.

Also it changes the signature of the directory filler function, adding a
flags parameter that we don't use anyway, so add wrappers to the readdir
function, and a compat helper for calling the filler function.

This change takes the code back to targeting FUSE2, but allows switching
to FUSE3 by adding -DFUSE3 to the CFLAGS and changing which version of
the library is linked to.